requirements.txt
fastapi
uvicorn
streamlit
requests
beautifulsoup4
sentence-transformers
faiss-cpu
pandas
openpyxl
numpy